requirements.txt
flask
huggingface_hub
pyngrok
python-dotenv
